[FEAT][Agent][api.py][v1/models]
